requirements.txt
ezdxf>=1.1.0
pydantic>=2.5.0
pyyaml>=6.0.1
tqdm>=4.66.0
streamlit>=1.30.0
openpyxl>=3.1.0
PySide6>=6.6.0
matplotlib>=3.8.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
jinja2>=3.1.0
numpy>=1.26.0
